refactor(router): mover placeholders de rutas públicas fuera de feature home
